fix: test code error and add embedded kafka redis
