+++ backend/src/test/java/com/dragonguard/backend/blockchain/controller/BlockchainControllerTest.java
@@ -4,17 +4,12 @@ import com.dragonguard.backend.blockchain.dto.response.BlockchainResponse;
 import com.dragonguard.backend.blockchain.entity.ContributeType;
 import com.dragonguard.backend.blockchain.service.BlockchainService;
 import com.dragonguard.backend.blockchain.service.TransactionService;
-import com.dragonguard.backend.commit.entity.Commit;
-import com.dragonguard.backend.member.entity.Member;
-import com.dragonguard.backend.member.repository.MemberRepository;
-import com.dragonguard.backend.member.service.AuthService;
-import com.dragonguard.backend.support.DatabaseTest;
+import com.dragonguard.backend.config.security.jwt.JwtTokenProvider;
+import com.dragonguard.backend.config.security.jwt.JwtValidator;
+import com.dragonguard.backend.config.security.oauth.user.UserDetailsMapper;
 import com.dragonguard.backend.support.docs.RestDocumentTest;
-import com.dragonguard.backend.support.fixture.member.entity.MemberFixture;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
@@ -26,33 +21,25 @@ import java.util.UUID;
 
 import static com.dragonguard.backend.support.docs.ApiDocumentUtils.getDocumentRequest;
 import static com.dragonguard.backend.support.docs.ApiDocumentUtils.getDocumentResponse;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@DatabaseTest
 @WebMvcTest(BlockchainController.class)
 class BlockchainControllerTest extends RestDocumentTest {
     @MockBean
     private BlockchainService blockchainService;
     @MockBean
-    protected AuthService authService;
-    @Autowired
-    private MemberRepository memberRepository;
-    protected Member loginUser;
-
-    @BeforeEach
-    public void setup() {
-        Member member = MemberFixture.SAMPLE1.toMember();
-        loginUser = memberRepository.save(member);
-        when(authService.getLoginUser()).thenReturn(loginUser);
-    }
+    private TransactionService transactionService;
+    @MockBean
+    private JwtValidator jwtValidator;
+    @MockBean
+    private JwtTokenProvider jwtTokenProvider;
+    @MockBean
+    private UserDetailsMapper userDetailsMapper;
 
     @Test
     @DisplayName("블록체인 부여 기록 리스트 조회")
@@ -64,7 +51,7 @@ class BlockchainControllerTest extends RestDocumentTest {
 
         ResultActions perform =
                 mockMvc.perform(
-                        get("/blockchain/1")
+                        get("/blockchain")
                                 .contentType(MediaType.APPLICATION_JSON));
 
         perform.andExpect(status().isOk())
