+++ backend/src/test/java/com/dragonguard/backend/search/controller/SearchControllerTest.java
@@ -1,19 +1,13 @@
 package com.dragonguard.backend.search.controller;
 
-import com.dragonguard.backend.commit.entity.Commit;
-import com.dragonguard.backend.member.entity.Member;
-import com.dragonguard.backend.member.entity.Role;
-import com.dragonguard.backend.member.repository.MemberRepository;
-import com.dragonguard.backend.member.service.AuthService;
+import com.dragonguard.backend.config.security.jwt.JwtTokenProvider;
+import com.dragonguard.backend.config.security.jwt.JwtValidator;
+import com.dragonguard.backend.config.security.oauth.user.UserDetailsMapper;
 import com.dragonguard.backend.result.dto.response.ResultResponse;
 import com.dragonguard.backend.search.service.SearchService;
-import com.dragonguard.backend.support.DatabaseTest;
 import com.dragonguard.backend.support.docs.RestDocumentTest;
-import com.dragonguard.backend.support.fixture.member.entity.MemberFixture;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
@@ -26,31 +20,22 @@ import static com.dragonguard.backend.support.docs.ApiDocumentUtils.getDocumentR
 import static com.dragonguard.backend.support.docs.ApiDocumentUtils.getDocumentResponse;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@DatabaseTest
 @WebMvcTest(SearchController.class)
 class SearchControllerTest extends RestDocumentTest {
     @MockBean
     private SearchService searchService;
     @MockBean
-    protected AuthService authService;
-    @Autowired
-    private MemberRepository memberRepository;
-    protected Member loginUser;
-
-    @BeforeEach
-    public void setup() {
-        Member member = MemberFixture.SAMPLE1.toMember();
-        loginUser = memberRepository.save(member);
-        when(authService.getLoginUser()).thenReturn(loginUser);
-    }
-
+    private JwtValidator jwtValidator;
+    @MockBean
+    private JwtTokenProvider jwtTokenProvider;
+    @MockBean
+    private UserDetailsMapper userDetailsMapper;
 
     @Test
     @DisplayName("검색 결과 조회")
